Removed redundant functions from clusters.py
