+++ app/main/controllers/__init__.py
@@ -1,16 +1,10 @@
 import os
 
-from flask import jsonify, abort, send_from_directory
-from ..models.models import Project, Source, Cluster, Item
+from flask import jsonify, send_from_directory
 from .projects import set_project_routes
 from .sources import set_source_routes
 from .connections import set_connection_routes
 from .clusters import set_cluster_routes
-from ..auth import requires_auth, AuthError
-
-
-
-
 
 
 def set_routes(app):
